Feat: Log in as admin and log in as user
