--- src/app/shared/auth/auth.component.html
@@ -1,5 +1,0 @@
-<p>
-  <button (click)="loginWithEmailAndPassword()" *ngIf="showLoginButton" >Log in</button>
-  <button (click)="loginWithGoogle()" *ngIf="showLoginButton" >Log in with Google</button>
-  <button (click)="logout()" *ngIf="showLogoutButton">Log out</button>
-</p>
