--- src/app/shared/auth/auth.component.ts
@@ -1,53 +1,0 @@
-import { Component, OnDestroy, OnInit, Optional } from '@angular/core';
-import { Auth, authState, signOut, User, GoogleAuthProvider, signInWithPopup, signInWithEmailAndPassword } from '@angular/fire/auth';
-import { EMPTY, Observable, Subscription } from 'rxjs';
-import { map } from 'rxjs/operators';
-import { traceUntilFirst } from '@angular/fire/performance';
-
-@Component({
-  selector: 'app-auth',
-  templateUrl: './auth.component.html',
-  styleUrls: ['./auth.component.scss']
-})
-export class AuthComponent implements OnInit, OnDestroy {
-
-  private readonly userDisposable: Subscription|undefined;
-  public readonly user: Observable<User | null> = EMPTY;
-
-  showLoginButton = true;
-  showLogoutButton = true;
-
-  constructor(@Optional() private auth: Auth) {
-    if (auth) {
-      this.user = authState(this.auth);
-      this.userDisposable = authState(this.auth).pipe(
-        traceUntilFirst('auth'),
-        map(u => !!u)
-      ).subscribe(isLoggedIn => {
-        this.showLoginButton = !isLoggedIn;
-        this.showLogoutButton = isLoggedIn;
-      });
-    }
-  }
-
-  ngOnInit(): void { }
-
-  ngOnDestroy(): void {
-    if (this.userDisposable) {
-      this.userDisposable.unsubscribe();
-    }
-  }
-
-  async loginWithEmailAndPassword() {
-    return await signInWithEmailAndPassword(this.auth, this.user, this.password);
-  }
-
-  async loginWithGoogle() {
-    return await signInWithPopup(this.auth, new GoogleAuthProvider());
-  }
-
-  async logout() {
-    return await signOut(this.auth);
-  }
-
-}
